pyproject.toml
[project]
name = "root-optimization"
version = "0.1.0"
description = ""
authors = [
    {name = "leanleanlean96",email = "kiraaff2289@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "geoalchemy2 (>=0.18.4,<0.19.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "shapely (>=2.1.2,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.7"

